Add files via upload
Found the best regression algorithm to predict customer lifetime value for car insurance companies. Algorithms with hyperparameters are implemented with GridSearchCV to find the best hyperparameter for that algorithm. Regression Algorithms used in this project are:
1. Linear Regression
2. Polynomial Linear Regression
3. Ridge
4. Lasso
5. Support Vector Machines (Simple and with kernels (linear, poly, rbf)
6. DecisionTreeRegressor
7. KnnRegressor
8. SGDRegressor

Note: This project is a collaborative effort. Paawan Bothra (Me) and Vijith Vengsarkar partnered to complete the project